Tensorflow_docker_deploy/requirement.txt
joblib
Flask
Numpy
Pandas
scikit-learn
tensorflow
opencv-python
paho-mqtt
python-etcd
Pillow
matplotlib
